+++ src/common/constants.ts
@@ -1,1 +1,2 @@
-export const ERROR_BOUNDARY_ROUTE_PATH = "/dev-error-boundary";
+export const ERROR_BOUNDARY_ROUTE_PATH_SOURCE = "/dev-error-boundary/source";
+export const ERROR_BOUNDARY_ERROR_CONTEXT = "/dev-error-boundary/error-context";