+++ src/common/simpleHash.ts
@@ -1,0 +1,9 @@
+export function simpleHash(str: string) {
+  let hash = 5381;
+  for (let i = 0; i < str.length; i++) {
+    hash = (hash << 5) - hash + str.charCodeAt(i);
+    hash |= 0;
+  }
+
+  return hash >>> 0;
+}